Enable GlslShaderGenerator to add layout/binding information in generated code. This makes the MaterialX Codegen'ed GLSL shader more SPIR-V friendly and also Cross Compiler friendly. However a side effect is that uniform initialization in the GLSL shader is not possible.
Additional details:
1. Separate and group uniforms:
The Hardware codegen will separate and group uniforms into three areas:
Private, Public and Sampler blocks
Public uniforms are inputs that should be published in a user interface for user interaction.
Private uniforms are internal variables needed by the system which should not be exposed in UI.
All texture based objects should be added to Sampler block

2. Emit uniform blocks with binding information
A new HwResourceBindingContext Interface and an implementation for GlslResourceBindingContext is added to allow GlslShaderGenerator to emit extended layout binding information where configured.

The configuration is HwResourceBindingContext is added to the Context by the caller as userdata called "udbinding" (HW::USER_DATA_BINDING_CONTEXT)