Add a helper to ensure u_envRadianceSamples has boounds.
PrivateUniforms_pixel translate to HLSL as
cbuffer PrivateUniforms_pixel : register(b2, space0)
{
    row_major float4x4 _2127_u_envMatrix : packoffset(c0);
    int _2127_u_envRadianceMips : packoffset(c4);
    int _2127_u_envRadianceSamples : packoffset(c4.y);
    float3 _2127_u_viewPosition : packoffset(c5);
    int _2127_u_numActiveLightSources : packoffset(c5.w);
};

_2127_u_envRadianceSamples causes trouble as it's used in a for loop
for (int i = 0; i < _2127_u_envRadianceSamples; i++)
  {...}

This shader causes a Driver Timeout (TDR) and seems to be something related to loop unroll

With _2127_u_envRadianceSamples as int, it tries to unroll and takes too long, hence TDRs.
Strangely, if we change _2127_u_envRadianceSamples to float, it does not unroll and runs the loop as-is

To keep it safe, a helper method is added ensure u_envRadianceSamples has upper-bound.