Update drawScreenSpaceQuad to accept optional Attibute id
The drawScreenSpaceQuad assumes
i_position at 0 and i_texcoord_0 at 1

But on macOS/GL Core Profile the locations are different.
e.g: i_position 1 and i_texcoord_0 0

This change queries the program for the input location.

Additonal changes include a helper in SimpleWindow to query scale factor for HDPI
(not currently used at the moment)