AutoTextureResolution option for Texture Baking.
When Enabled:
If any images are found  upstream from a shader input, then the baked
output Image size is set to the largest image resolution automatically.
If no images are found, then the fixed resolution of the baker is used.

When Disabled:
Image on shader input are baked to fixed resolution of the baker.