Initial commit legacy materials
This commit includes:
 - all Legacy (Protein) classes
 - procedural nodes (checker, noise, turbulence 2d/3d)
 - utilities (stain, normals combine nodes)

For now all nodes are under "adsk_legacy" group, except the turbulence ones that might be ok to be added to the standard lib.

Becasue of a bug in the handling of namespaces in the Graph Editor all namespaces are set to "". Bug: https://github.com/AcademySoftwareFoundation/MaterialX/issues/1695