USD test file
This is a USD test file that can be loaded in usdview to test the same materials that the legacy_materials_test.mtlx tests in MaterialX viewer.

Open the file in usdview, select the ShaderBall object and in the  Meta data panel you can select a variant for each material class.

The legacy materials node defs are also updated to remove a stub for namespace that was causing trouble in usdview (but was ok in MaterialX viewer).
A typo in the comments was also fixed.