Re-Adding degrees for Noise
Apparently, unit="degree" is still needed for Noise (and Noise only).
This is a bug in Materialx (active in the issues list) and I still don't know when the unit is needed not not.
Technically, it should not be there, but in this case it throws an error (he's trying to convert to "meters" for some reason).