Coordinates inputs revision, enums, and other
1) Changing all separate inputs to vectors (offsets 2d and 3d, rotation 3d)
Unified names to match Prism:
realworld_offset (vector2 or vector3)
realworld_scale (vector2)
rotation_angle (float or vector3)

2) Change inputs secting behaviours to be integer and adding enum and enumvalues attributes.
This affects Noise and Gradient Types in one or more places.
Note that for now the integer input is converted to float inside the graph to use float math.
This will be replaced in the future with integer math now that we have switched to MaterialX 1.39, where it's finally available.

3) Set all texcoord and position inputs to uivisible="false" (except util nodes)

4) A few inputs were renamed to remove underscore (noise and marble)
That's ok for now, and make most inputs consistent, but in a next PR I will instead unify all inputs to use underscores as separators.