input renames
Unify the procedurals input names format to use underscores like Prism and Protein materials.

Most changes in the nodedefs are simply underscores added to separate words.
In a very few cases, the rename was more drastic to clarify the input meaning.

The nodegraphs reflects the changes to the nodedefs, but in some cases a few nodes inside had to be renamed because they had the same name as the new inputs, and that is not allowed.