Adding legacy knurl
Adding nodedef and nodegraph for Legacy Knurl.
This implements a generic knurl that is a superset, usable standalone, of the knurl used as a Metal pattern.
The original knurl in RapidRT was not usable standalone. This solution is different and has more features, making it a worthwhile extra procedural to add to the library.

A coupe of extra fixes are included
- defaults for a couple of size parameters were incorrectly set to 0,0 after changing them to vector2.
- A converter node from LookdevX was removed. It was unconnected, so no real change, and it might be recreated if that graph is re-exported. I'll try to keep an eye on it.