Adding diagonal option
The diagonal option was in my original version, but was removed when I added separate X and Y scales, as the pattern was distorting and it was not necessary for matching the RRT version.

But then I realized this was an important feature of Knurl in manufacturing, and spend some time trying to make it work again.

This change adds just a few nodes and the feature is back, even if there is a small approximation in the normals direction when stretched.