Adding input to hardwood
Adding an input to Hardwood to control amount of the automatic pattern (based on wood grain)

This is used by the next round of updates for the materials converter.

Note the default amount is low, and based on visual estimates, but it's still a bit high. This is an issue captured in Jira, where all bump amounts seem to be extremely high.