Enabling flakes in Metallicpaint graph
This adds inputs related to flakes in the nodedef, and a bunch of nodes to handle flakes as alternative inputs to the standard surface material.

One note of caution, when flakes are enabled the Highlight Spread takes a different meaning. Nothing to do in the graph, but the converter need to be aware of it and change the default value.