Fix for mismatching nodedef
LookdevX seem to have xported a "divide" node using  a "multiply" nodedef. This caused a lot of issues and is something we might have to keep an eye for.

Here I re-exported the graph from LookdevX and this time it exported it correctly. But being a re-export after a few debugging attempts, a few other things are changed in the diff.
The only meaningful change is the "nodedef="in the  "divide" line.