Flakes procedural (#1453)
* Flakes procedural

Adding a procedural nodedef/graph for metallic paint flakes to be used in the converter.
This procedural will replaces whatever was in Protein/RapidRT, since they were not realistic and using custom noises procedurals.
This graph uses only existing procedurals and noises.
The flakes shapes is different, but likely more realistic. The overall look of the flakes effect is also different, as RapidRT (and previously, Mental Ray) didn't seem to have any actual metallic reflection.
Compared to real metallic paint references, this procedural produces a much more realistic result.

* Adjusting defaults and gamma

Going for a slightly stronger defaults.
Also changing the gamma curve to strengthen the fleck affect.

* Enabling flakes in Metallicpaint graph

This adds inputs related to flakes in the nodedef, and a bunch of nodes to handle flakes as alternative inputs to the standard surface material.

One note of caution, when flakes are enabled the Highlight Spread takes a different meaning. Nothing to do in the graph, but the converter need to be aware of it and change the default value.

* Fix for mismatching nodedef

LookdevX seem to have xported a "divide" node using  a "multiply" nodedef. This caused a lot of issues and is something we might have to keep an eye for.

Here I re-exported the graph from LookdevX and this time it exported it correctly. But being a re-export after a few debugging attempts, a few other things are changed in the diff.
The only meaningful change is the "nodedef="in the  "divide" line.

* Flakes off by default

Flakes should be off by default. They were erroneusly set to On.

* Adjustment to density default

Comparing converter and results on a one inch cube shows that the original default is a bit too dense.

Values like this will likely be tweaked multiple times as we keep working on it.