legacy_glass phase3
Experimental node with embedded relief procedurals.

- Added procedurals for Rippled and Wavy reliefs.
- Added an Amount control to set the weight of the procedural reliefs
  - Note that the Amount control also affects the Custom relief input
- Added a texcoord input for the embedded reliefs. This does not affect Custom relief.

- Added a relief_scale input with units. This allows control of units for both the procedurals and the matching heighttonormal. This is important as the texcoord need to match everywhere and default values do not allow so.
There is an extra benefit that the scale could actually be useful for a user to adjust the overall sizes, so it was made visible in the UI.
Additional benefit (and major win!): with this we do not need to manually add unittype and unit in the utility nodes inside the graph where units would have been necessary, avoiding manual changes every time after publishing the graphs from LookdevX

Possible changes:
 - texcoord rename to relief_texcoord? It's only for the embedded relief presets, not for other input textures color or even the custom relief. Not sure if we need to make this evident in the input name or separately in some documentations

Note that normal inputs for rippled and wavy are still present but disconnected. I will remove them once this experiment has proven valid, and does not need to revert to the previous converter style where graphs are created by code outside the material graph.