legacy_stone phase3 (#1460)
* legacy_stone phase3

Adding preset finish bump support
This is based on internal adsk:height_map nodes which have all the proper defaults except the filename which will need to be passed on by the converter.

This is an experimental graph (like for Glass). The handling of the sizes is done trough the extra (but somewhat useful) finish_scale input with is unit based. This avoids having to use units in the adsk:height_map nodes, which would need to be added manually.

Note that  the adsk: namespace causes trouble in Graph Editor. Namespaced nodes can't be created there. In LookdevX they work, but they export without namespace in the nodes. This requires manual fixes, but then they work in Graph Editor and Viewer too.

* Swapping preset order

Matching the order of the Protein presets