Bump amount added (#1462)
Adding an input to control bump amount so it ca be exposed in the UI as done for Glass and Stone.
Note that there is an Amount from the adsk_height_map as well, and the two can be used together,

This is experimental and will need validation.

Also changed the IOR defaults to match Protein. These values are debatable but for the sake of consistency I decided to match them.